LayoutTests/imported/w3c/web-platform-tests/css/css-values/viewport-units-scrollbars-compute-expected.txt



FAIL 100vw computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100vi computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100vmax computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100svw computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100svi computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100svmax computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100lvw computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100lvi computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100lvmax computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100dvw computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100dvi computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100dvmax computes to 185px assert_equals: expected "185px" but got "200px"
FAIL 100vh computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100vb computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100vmin computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100svh computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100svb computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100svmin computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100lvh computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100lvb computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100lvmin computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100dvh computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100dvb computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 100dvmin computes to 85px assert_equals: expected "85px" but got "100px"
FAIL 1dvw computes to 1.84375px assert_equals: expected "1.84375px" but got "2px"
FAIL 10dvw computes to 18.5px assert_equals: expected "18.5px" but got "20px"
FAIL 1dvh computes to 0.84375px assert_equals: expected "0.84375px" but got "1px"
FAIL 10dvh computes to 8.5px assert_equals: expected "8.5px" but got "10px"
FAIL calc(1dvw + 1dvw) computes to 3.6875px assert_equals: expected "3.6875px" but got "4px"
FAIL calc(1dvw + 1dvh) computes to 2.6875px assert_equals: expected "2.6875px" but got "3px"
FAIL calc(1dvw + 100px) computes to 101.844px assert_equals: expected "101.844px" but got "102px"
FAIL max(1svw, 1svh) computes to 1.84375px assert_equals: expected "1.84375px" but got "2px"
FAIL min(1lvw, 1lvh) computes to 0.84375px assert_equals: expected "0.84375px" but got "1px"
FAIL calc(1dvw + 10%) computes to 10.3438px assert_equals: expected "10.3438px" but got "10.5px"
